KB-9962 | BE|DEV| Flink Job enhancements to update the certificate count to the redis
1. Certificate count enhancements.
