+++ credential-generator/event-cert-generator/src/main/resources/event-certificate-generator.conf
@@ -40,3 +40,10 @@ enable.suppress.exception = true
 enable.rc.certificate = false
 minCompletePercentageForCert=50
 specialEventCertificateName = "RepublicDay_2025"
+
+
+redis {
+  database {
+   certificateCountCache.id = 0
+  }
+}
