+++ credential-generator/event-cert-generator/src/main/scala/org/event/job/certgen/function/CertificateGeneratorFunction.scala
@@ -28,6 +28,7 @@ import java.util.stream.Collectors
 import scala.collection.JavaConverters.mapAsJavaMap
 import org.event.job.certgen.domain.{Actor, BEJobRequestEvent, Certificate, CertificateAuditEvent, Event, EventContext, EventObject, EventObjectCourseCertificate, Issuer, Recipient, Training, UserEnrollmentData}
 import org.event.job.certgen.exception.ServerException
+import org.sunbird.job.cache.{DataCache, RedisConnect}
 
 import scala.collection.JavaConverters._
 
@@ -42,12 +43,16 @@ class CertificateGeneratorFunction  (config: EventCertificateGeneratorConfig, ht
   implicit val certificateConfig: CertificateConfig = CertificateConfig(basePath = config.basePath, encryptionServiceUrl = config.encServiceUrl, contextUrl = config.CONTEXT, issuerUrl = config.ISSUER_URL,
     evidenceUrl = config.EVIDENCE_URL, signatoryExtension = config.SIGNATORY_EXTENSION)
   implicit var esUtil: ElasticSearchUtil = null
+  private var dataCache: DataCache = _
 
   override def open(parameters: Configuration): Unit = {
     super.open(parameters)
     cassandraUtil = new CassandraUtil(config.dbHost, config.dbPort)
     if (esUtil == null)
       esUtil = new ElasticSearchUtil(config.esConnection, config.certIndex, "config.auditHistoryIndexType")
+    val redisConnect = new RedisConnect(config)
+    dataCache = new DataCache(config, redisConnect, config.cacheDbId, List())
+    dataCache.init()
   }
 
   override def close(): Unit = {
@@ -98,6 +103,10 @@ class CertificateGeneratorFunction  (config: EventCertificateGeneratorConfig, ht
     val certModelList: List[CertModel] = new CertMapper(certificateConfig).mapReqToCertModel(event)
     println("generateCertificate " + event)
     val certificateGenerator = new CertificateGenerator
+    val primaryFields = Map(config.userId.toLowerCase() -> event.userId,
+      config.batchId.toLowerCase -> event.batchId,
+      config.courseId.toLowerCase -> event.courseId)
+    val increaseCertCount: Boolean = getIssuedCertificatesDetailsFromUserEnrollmentTable(primaryFields)
     certModelList.foreach(certModel => {
       var uuid: String = null
       try {
@@ -131,6 +140,18 @@ class CertificateGeneratorFunction  (config: EventCertificateGeneratorConfig, ht
         cleanUp(uuid, directory)
       }
     })
+    if (increaseCertCount) {
+      val redisKey = s"user:certCount:$event.userId"
+      val redisValue = 1.toString
+      val redisUserCertificateCount = dataCache.getStringValue(redisKey)
+      if (redisUserCertificateCount.nonEmpty) {
+        val updatedRedisValue = redisUserCertificateCount.toInt + 1
+        dataCache.setWithRetryAndTTL(redisKey, updatedRedisValue.toString)
+      } else {
+        dataCache.setWithRetryAndTTL(redisKey, redisValue)
+      }
+    }
+
   }
 
   @throws[Exception]
@@ -452,5 +473,29 @@ class CertificateGeneratorFunction  (config: EventCertificateGeneratorConfig, ht
       "object" -> Map("id" -> event.userId, "type" -> "EventEnrolmentAlert")
     )
     ScalaJsonUtil.serialize(message)
+  }
+
+  private def getIssuedCertificatesDetailsFromUserEnrollmentTable(columns: Map[String, AnyRef]): Boolean = {
+    logger.info("primary columns {}", columns)
+    val selectWhere = QueryBuilder.select("issued_certificates", "status", "progress")
+      .from(config.dbKeyspace, config.dbEnrollmentTable)
+      .where()
+    columns.map(col => {
+      col._2 match {
+        case value: List[Any] =>
+          selectWhere.and(QueryBuilder.in(col._1, value.asJava))
+        case _ =>
+          selectWhere.and(QueryBuilder.eq(col._1, col._2))
+      }
+    })
+    logger.info("select query {}", selectWhere.toString)
+    val records = cassandraUtil.find(selectWhere.toString).asScala.toList
+    //Check status == 2 and issued_certificates is not empty if so, return false -- means increaseCertCount is false
+    !records.exists(row => {
+      val certificates = row.getObject("issued_certificates")
+        .asInstanceOf[java.util.List[java.util.Map[String, String]]]
+      val status = row.getInt("status")
+      certificates != null && !certificates.isEmpty && status == 2
+    })
   }
 }
